dataflow_beam_v0/dataflow-beam-raw/requirements.txt
kafka-python==2.0.2
pandas
beam_nuggets
fsspec
gcsfs